backend/requirements.txt
Flask
flask-cors
firebase-admin
gunicorn
python-dotenv
firebase-admin
pytz
